Fix path collision when using context (#27)
In v1.5.0, if `context` is provided without `dockerfile`, it would still
run with `--file Dockerfile`, which does NOT match the default behavior
of a Docker build. This tweaks the build args to only pass the `--file`
argument if it was explicitly provided. Consequently, builds of a
subdirectory (etc) w/out giving a path to the dockerfile should work by
default.